backend/requirements.txt
fastapi
uvicorn
pydantic
asyncpg
python-jose[cryptography]  # For JWT
passlib[bcrypt]  # For password hashing
python-multipart
aioredis